voter_project/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cryptography==46.0.3
cssselect2==0.8.0
diff-match-patch==20241021
Django==4.2.26
django-import-export==4.3.14
django-restframework==0.0.1
django-seed==0.3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
Faker==37.12.0
fastapi==0.121.3
fonttools==4.60.1
frozenlist==1.8.0
geoip2==5.1.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
maxminddb==2.8.2
multidict==6.7.0
numpy==2.0.2
opencv-python==4.12.0.88
pdfkit==1.0.0
pillow==11.3.0
propcache==0.4.1
psycopg2==2.9.11
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
requests==2.32.5
rfc3986==1.5.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.49.3
tablib==3.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toposort==1.10
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
weasyprint==66.0
webencodings==0.5.1
yarl==1.22.0
zopfli==0.2.3.post1
